refactor: clippy idiomatic cleanups + fix stale 45->54 tool-count docs
Applied 16 machine-verified clippy fixes across 11 memory modules
(Error::other, slice trim_end_matches, contains(), is_some_and,
repeat_n, Range::contains, collapsed ifs, dropped a redundant f32
cast in the reranker score path). Behaviour-preserving: full suite
still green (133 tests). Fixed the registry module docs + a research
doc that still called it a '45-tool surface' — it is 54. Clippy lib
warnings 31 -> 15 (remainder are subjective type/arg-count refactors).

+++ src-tauri/src/memory/reranker.rs
@@ -80,7 +80,7 @@ pub fn rerank(query: &str, documents: &[String]) -> Result<Vec<f32>> {
     let mut by_idx: Vec<f32> = vec![0.0; documents.len()];
     for r in results {
         if r.index < by_idx.len() {
-            by_idx[r.index] = r.score as f32;
+            by_idx[r.index] = r.score;
         }
     }
     Ok(by_idx)
